new `session_state.util` module
