+++ lib/streamlit/runtime/state/util.py
@@ -1,0 +1,51 @@
+# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import hashlib
+from typing import Optional
+
+from streamlit.runtime.state.session_state import GENERATED_WIDGET_KEY_PREFIX
+from streamlit.runtime.state.widgets import WidgetProto
+
+
+def compute_widget_id(
+    element_type: str, element_proto: WidgetProto, user_key: Optional[str] = None
+) -> str:
+    """Compute the widget id for the given widget. This id is stable: a given
+    set of inputs to this function will always produce the same widget id output.
+
+    The widget id includes the user_key so widgets with identical arguments can
+    use it to be distinct.
+
+    The widget id includes an easily identified prefix, and the user_key as a
+    suffix, to make it easy to identify it and know if a key maps to it.
+
+    Does not mutate the element_proto object.
+    """
+    h = hashlib.new("md5")
+    h.update(element_type.encode("utf-8"))
+    h.update(element_proto.SerializeToString())
+    return f"{GENERATED_WIDGET_KEY_PREFIX}-{h.hexdigest()}-{user_key}"
+
+
+def user_key_from_widget_id(widget_id: str) -> Optional[str]:
+    """Return the user key portion of a widget id, or None if the id does not
+    have a user key.
+
+    TODO This will incorrectly indicate no user key if the user actually provides
+    "None" as a key, but we can't avoid this kind of problem while storing the
+    string representation of the no-user-key sentinel as part of the widget id.
+    """
+    user_key = widget_id.split("-", maxsplit=2)[-1]
+    user_key = None if user_key == "None" else user_key
+    return user_key
